ndb fix windows warning

+++ storage/ndb/src/ndbapi/trp_buffer.hpp
@@ -129,7 +129,7 @@ public:
   TFPool();
   ~TFPool();
 
-  bool init(Uint64 total_memory);
+  bool init(size_t total_memory);
 
   TFPage* try_alloc(Uint32 N); // Return linked list of most N pages
   Uint32 try_alloc(struct iovec tmp[], Uint32 cnt);
